+++ pkg/db/ent/loginhistory/loginhistory.go
@@ -47,7 +47,7 @@ func ValidColumn(column string) bool {
 
 var (
 	// DefaultCreateAt holds the default value on creation for the "create_at" field.
-	DefaultCreateAt func() int64
+	DefaultCreateAt func() uint32
 	// DefaultID holds the default value on creation for the "id" field.
 	DefaultID func() uuid.UUID
 )
