+++ pkg/db/ent/loginhistory/where.go
@@ -120,7 +120,7 @@ func UserAgent(v string) predicate.LoginHistory {
 }
 
 // CreateAt applies equality check predicate on the "create_at" field. It's identical to CreateAtEQ.
-func CreateAt(v int64) predicate.LoginHistory {
+func CreateAt(v uint32) predicate.LoginHistory {
 	return predicate.LoginHistory(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldCreateAt), v))
 	})
@@ -501,21 +501,21 @@ func UserAgentContainsFold(v string) predicate.LoginHistory {
 }
 
 // CreateAtEQ applies the EQ predicate on the "create_at" field.
-func CreateAtEQ(v int64) predicate.LoginHistory {
+func CreateAtEQ(v uint32) predicate.LoginHistory {
 	return predicate.LoginHistory(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldCreateAt), v))
 	})
 }
 
 // CreateAtNEQ applies the NEQ predicate on the "create_at" field.
-func CreateAtNEQ(v int64) predicate.LoginHistory {
+func CreateAtNEQ(v uint32) predicate.LoginHistory {
 	return predicate.LoginHistory(func(s *sql.Selector) {
 		s.Where(sql.NEQ(s.C(FieldCreateAt), v))
 	})
 }
 
 // CreateAtIn applies the In predicate on the "create_at" field.
-func CreateAtIn(vs ...int64) predicate.LoginHistory {
+func CreateAtIn(vs ...uint32) predicate.LoginHistory {
 	v := make([]interface{}, len(vs))
 	for i := range v {
 		v[i] = vs[i]
@@ -532,7 +532,7 @@ func CreateAtIn(vs ...int64) predicate.LoginHistory {
 }
 
 // CreateAtNotIn applies the NotIn predicate on the "create_at" field.
-func CreateAtNotIn(vs ...int64) predicate.LoginHistory {
+func CreateAtNotIn(vs ...uint32) predicate.LoginHistory {
 	v := make([]interface{}, len(vs))
 	for i := range v {
 		v[i] = vs[i]
@@ -549,28 +549,28 @@ func CreateAtNotIn(vs ...int64) predicate.LoginHistory {
 }
 
 // CreateAtGT applies the GT predicate on the "create_at" field.
-func CreateAtGT(v int64) predicate.LoginHistory {
+func CreateAtGT(v uint32) predicate.LoginHistory {
 	return predicate.LoginHistory(func(s *sql.Selector) {
 		s.Where(sql.GT(s.C(FieldCreateAt), v))
 	})
 }
 
 // CreateAtGTE applies the GTE predicate on the "create_at" field.
-func CreateAtGTE(v int64) predicate.LoginHistory {
+func CreateAtGTE(v uint32) predicate.LoginHistory {
 	return predicate.LoginHistory(func(s *sql.Selector) {
 		s.Where(sql.GTE(s.C(FieldCreateAt), v))
 	})
 }
 
 // CreateAtLT applies the LT predicate on the "create_at" field.
-func CreateAtLT(v int64) predicate.LoginHistory {
+func CreateAtLT(v uint32) predicate.LoginHistory {
 	return predicate.LoginHistory(func(s *sql.Selector) {
 		s.Where(sql.LT(s.C(FieldCreateAt), v))
 	})
 }
 
 // CreateAtLTE applies the LTE predicate on the "create_at" field.
-func CreateAtLTE(v int64) predicate.LoginHistory {
+func CreateAtLTE(v uint32) predicate.LoginHistory {
 	return predicate.LoginHistory(func(s *sql.Selector) {
 		s.Where(sql.LTE(s.C(FieldCreateAt), v))
 	})
